+++ Modules~/AssetDeliveryModule/Scripts/Runtime/AssemblyInfo.cs
@@ -1,3 +1,3 @@
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("FlowIoC.Dev.Editor.Tests")]
+[assembly: InternalsVisibleTo("FlowIoC.Dev.Editor.ModuleTests")]