+++ Runtime/AssemblyInfo.cs
@@ -4,4 +4,4 @@ using UnityEngine.Scripting;
 [assembly: Preserve]
 [assembly: AlwaysLinkAssembly]
 [assembly: InternalsVisibleTo("FlowIoC.Editor")]
-[assembly: InternalsVisibleTo("FlowIoC.Tests")]
+[assembly: InternalsVisibleTo("FlowIoC.Dev.Editor.CoreTests")]